refactoring: move institution-group-util.ts
